backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-translate==3.12.0
google-cloud-speech==2.21.0
pydantic==2.4.2
python-multipart==0.0.6
websockets==11.0.3 